DIXON EDIT
Files in here allow for setting a default skin. If there is ever a issue
with your character's model in game, it will default back to this. It
will also use this model once you die and respawn. In addition, changes
here allow the use of ANY classname in arma currently, regardless of
gear. Example, before delta force would give you GPS and other gear
because it was a re-done classname. Now it does not give extra gear and
only applies the model to your character.